New dashboard replacing both old versions
Noah just created a new dashboard from the json located on the :8877/metrics endpoint doc. Process for adding the dashboard is the same. The dashboard with ID: 4698 has been moved to the bottom of additional install options since it supports metrics scraped from StatsD. The dashboard hopefully shouldn't need to be changed from here on out.